First version of experiment. Started updating algorithms to use AStar implementation of alignment.
git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/SelectivePRepair/Trunk@16619 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/jbpt/mining/repair/RepairConstraint.java
@@ -3,25 +3,25 @@ package org.jbpt.mining.repair;
 import java.util.Map;
 
 public class RepairConstraint {
-	private Map<String,Double> skipCosts = null;
-	private Map<String,Double> insertCosts = null;
-	private double resources = 0.0;
+	private Map<String,Integer> skipCosts = null;
+	private Map<String,Integer> insertCosts = null;
+	private int resources = 0;
 	
-	public RepairConstraint(Map<String,Double> insertCosts, Map<String,Double> skipCosts, double resources) {
+	public RepairConstraint(Map<String,Integer> insertCosts, Map<String,Integer> skipCosts, int resources) {
 		this.insertCosts = insertCosts;
 		this.skipCosts = skipCosts;
 		this.resources = resources;
 	}
 	
-	public Map<String, Double> getSkipCosts() {
+	public Map<String,Integer> getSkipCosts() {
 		return skipCosts;
 	}
 
-	public Map<String, Double> getInsertCosts() {
+	public Map<String,Integer> getInsertCosts() {
 		return insertCosts;
 	}
 
-	public double getAvailableResources() {
+	public int getAvailableResources() {
 		return resources;
 	}
 }
